fixed ShiftDAO

+++ src/main/java/iuh/fit/dao/ShiftDAO.java
@@ -95,7 +95,7 @@ public class ShiftDAO {
             preparedStatement.setTime(2, ConvertHelper.timeConvertertoSQL(shift.getStartTime()));
             preparedStatement.setTime(3, ConvertHelper.timeConvertertoSQL(shift.getEndTime()));
             preparedStatement.setTimestamp(4, ConvertHelper.dateTimeToSQLConverter(shift.getUpdatedDate()));
-            preparedStatement.setInt(5, shift.getNumberOfHour());
+            preparedStatement.setDouble(5, shift.getNumberOfHour());
             preparedStatement.setString(6, ConvertHelper.shiftDaysScheduleConverterToSQL(shift.getShiftDaysSchedule()));
 
             preparedStatement.executeUpdate();
@@ -133,7 +133,7 @@ public class ShiftDAO {
             preparedStatement.setTime(1, ConvertHelper.timeConvertertoSQL(shift.getStartTime()));
             preparedStatement.setTime(2, ConvertHelper.timeConvertertoSQL(shift.getEndTime()));
             preparedStatement.setTimestamp(3, ConvertHelper.dateTimeToSQLConverter(shift.getUpdatedDate()));
-            preparedStatement.setInt(4, shift.getNumberOfHour());
+            preparedStatement.setDouble(4, shift.getNumberOfHour());
             preparedStatement.setString(5, ConvertHelper.shiftDaysScheduleConverterToSQL(shift.getShiftDaysSchedule()));
             preparedStatement.setString(6, shift.getShiftID());
